requirements.txt
fastapi
uvicorn
python-dotenv
pydantic
httpx
